Added reusable component

+++ app/controllers/polls_controller.rb
@@ -9,6 +9,7 @@ class PollsController < ApplicationController
 
     def create
         @poll = Poll.new(poll_params.merge(creator_id: @current_user.id))
+        authorize @poll
         if @poll.save
             render status: :ok, json: { notice:  t('successfully_created', entity: 'Poll') }
         else 
@@ -18,11 +19,13 @@ class PollsController < ApplicationController
     end
 
     def show
+        authorize @poll
         @options = @poll.options
         render status: :ok, json: { poll: @poll, options: @options}
     end
 
     def update
+        authorize@poll
         if @poll.update(poll_params)
             render status: :ok, json: { notice: 'Successfully updated poll.' }
         else
@@ -31,6 +34,7 @@ class PollsController < ApplicationController
     end
 
     def destroy
+        authorize @task
         if @poll.destroy
             render status: :ok, json: { notice: 'Successfully deleted poll.' }
         else
